Moved the introspection interpreter
